// src/app/shared/model/AuctionSort.ts
export enum AuctionSort {
  ID_ASC = 'ID_ASC',
  ID_DESC = 'ID_DESC',
  BID_PRICE_ASC = 'BID_PRICE_ASC',
  BID_PRICE_DESC = 'BID_PRICE_DESC',
  BUY_NOW_PRICE_ASC = 'BUY_NOW_PRICE_ASC',
  BUY_NOW_PRICE_DESC = 'BUY_NOW_PRICE_DESC',
  TITLE_ASC = 'TITLE_ASC',
  TITLE_DESC = 'TITLE_DESC',
  END_TIME_ASC = 'END_TIME_ASC',
  END_TIME_DESC = 'END_TIME_DESC'
}